# openapi.json
{"openapi": "3.0.2", "info": {"title": "ContentAPI", "description": "\nContentAPI helps you do awesome stuff. \ud83d\ude80\n", "version": "1.0.0"}, "paths": {"/v1/quizzes/categories": {"get": {"tags": ["Quizzes"], "summary": "Get Categories", "description": "Get the categories from the quiz api.", "operationId": "get_categories_v1_quizzes_categories_get", "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/CategoriesModel"}, "example": {"categories": ["Arts & Literature", "Film & TV"]}}}}, "500": {"description": "Something went wrong"}}}}, "/v1/quizzes/tags": {"get": {"tags": ["Quizzes"], "summary": "Get Tags", "description": "Get the tags from the quiz server.", "operationId": "get_tags_v1_quizzes_tags_get", "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/TagsModel"}, "example": {"tags": ["alcohol", "acting"]}}}}, "500": {"description": "Something went wrong"}}}}, "/v1/quizzes/": {"post": {"tags": ["Quizzes"], "summary": "Post User Quiz", "description": "Create a user quiz.", "operationId": "post_user_quiz_v1_quizzes__post", "security": [{"HTTPBearer": []}], "requestBody": {"required": true, "content": {"application/json": {"schema": {"$ref": "#/components/schemas/CreateUserQuiz"}}}}, "responses": {"201": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/UserQuiz"}, "example": {"category": "History", "difficulty": "easy", "questions": [{"correct_answer": "in London in 1912", "incorrect_answers": ["in Manchester in 1901", "in Oxford in 1924", "in Cambridge in 1935"], "question": "Where and when was Alan Turing born?"}, {"correct_answer": "a Turing Machine", "incorrect_answers": ["the Bombe", "Church's Computer", "the Manchester Mach I"], "question": "Turing, while solving the Decision Problem, proposed a hypothetical computing machine, which we now call ___"}], "tags": "1910's", "title": "The father of the computer", "uid": "XXXXXXXXXXXXXXXXXXXXXXXXXXXX"}}}}, "401": {"content": {"application/json": {"example": {"detail": "Bearer authentication is needed"}}}, "description": "Unauthorized"}, "404": {"content": {"application/json": {"example": {"detail": "User cannot be found"}}}, "description": "Not Found"}, "500": {"description": "Internal Server Error"}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}, "get": {"tags": ["Quizzes"], "summary": "Get User Quizzes", "description": "Get the user's quizzes from the database.", "operationId": "get_user_quizzes_v1_quizzes__get", "security": [{"HTTPBearer": []}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/UserQuizzes"}, "example": {"YYYYYYYYYYYYYYYYYYYY": {"category": "History", "difficulty": "easy", "questions": [{"correct_answer": "in London in 1912", "incorrect_answers": ["in Manchester in 1901", "in Oxford in 1924", "in Cambridge in 1935"], "question": "Where and when was Alan Turing born?"}, {"correct_answer": "a Turing Machine", "incorrect_answers": ["the Bombe", "Church's Computer", "the Manchester Mach I"], "question": "Turing, while solving the Decision Problem, proposed a hypothetical computing machine, which we now call ___"}], "tags": "1910's", "title": "The father of the computer", "uid": "XXXXXXXXXXXXXXXXXXXXXXXXXXXX"}}}}}, "401": {"content": {"application/json": {"example": {"detail": "Bearer authentication is needed"}}}, "description": "Unauthorized"}, "404": {"content": {"application/json": {"example": {"detail": "User cannot be found"}}}, "description": "Not Found"}, "500": {"description": "Internal Server Error"}}}, "delete": {"tags": ["Quizzes"], "summary": "Delete User Quizzes", "description": "Delete user quizzes.", "operationId": "delete_user_quizzes_v1_quizzes__delete", "security": [{"HTTPBearer": []}], "requestBody": {"required": true, "content": {"application/json": {"schema": {"$ref": "#/components/schemas/DeleteUserQuizzes"}}}}, "responses": {"204": {"description": "Successful Response"}, "401": {"content": {"application/json": {"example": {"detail": "Bearer authentication is needed"}}}, "description": "Unauthorized"}, "404": {"content": {"application/json": {"example": {"detail": "User cannot be found"}}}, "description": "Not Found"}, "500": {"description": "Internal Server Error"}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/v1/quizzes/{quiz_id}": {"get": {"tags": ["Quizzes"], "summary": "Get User Quiz", "description": "Get user quiz.", "operationId": "get_user_quiz_v1_quizzes__quiz_id__get", "security": [{"HTTPBearer": []}], "parameters": [{"name": "quiz_id", "in": "path", "required": true, "schema": {"title": "Quiz Id", "type": "string"}}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/UserQuiz"}, "example": {"category": "History", "difficulty": "easy", "questions": [{"correct_answer": "in London in 1912", "incorrect_answers": ["in Manchester in 1901", "in Oxford in 1924", "in Cambridge in 1935"], "question": "Where and when was Alan Turing born?"}, {"correct_answer": "a Turing Machine", "incorrect_answers": ["the Bombe", "Church's Computer", "the Manchester Mach I"], "question": "Turing, while solving the Decision Problem, proposed a hypothetical computing machine, which we now call ___"}], "tags": "1910's", "title": "The father of the computer", "uid": "XXXXXXXXXXXXXXXXXXXXXXXXXXXX"}}}}, "401": {"content": {"application/json": {"example": {"detail": "Bearer authentication is needed"}}}, "description": "Unauthorized"}, "404": {"content": {"application/json": {"example": {"detail": "User cannot be found"}}}, "description": "Not Found"}, "500": {"description": "Internal Server Error"}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}, "delete": {"tags": ["Quizzes"], "summary": "Delete User Quiz", "description": "Delete a user quiz.", "operationId": "delete_user_quiz_v1_quizzes__quiz_id__delete", "security": [{"HTTPBearer": []}], "parameters": [{"name": "quiz_id", "in": "path", "required": true, "schema": {"title": "Quiz Id", "type": "string"}}], "responses": {"204": {"description": "Successful Response"}, "401": {"content": {"application/json": {"example": {"detail": "Bearer authentication is needed"}}}, "description": "Unauthorized"}, "404": {"content": {"application/json": {"example": {"detail": "User cannot be found"}}}, "description": "Not Found"}, "500": {"description": "Internal Server Error"}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}, "patch": {"tags": ["Quizzes"], "summary": "Patch User Quiz", "description": "Update user quiz.", "operationId": "patch_user_quiz_v1_quizzes__quiz_id__patch", "security": [{"HTTPBearer": []}], "parameters": [{"name": "quiz_id", "in": "path", "required": true, "schema": {"title": "Quiz Id", "type": "string"}}], "requestBody": {"required": true, "content": {"application/json": {"schema": {"$ref": "#/components/schemas/UpdateUserQuiz"}}}}, "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}, "example": {"category": "History", "difficulty": "easy", "questions": [{"correct_answer": "in London in 1912", "incorrect_answers": ["in Manchester in 1901", "in Oxford in 1924", "in Cambridge in 1935"], "question": "Where and when was Alan Turing born?"}, {"correct_answer": "a Turing Machine", "incorrect_answers": ["the Bombe", "Church's Computer", "the Manchester Mach I"], "question": "Turing, while solving the Decision Problem, proposed a hypothetical computing machine, which we now call ___"}], "tags": "1910's", "title": "The father of the computer", "uid": "XXXXXXXXXXXXXXXXXXXXXXXXXXXX"}}}}, "401": {"content": {"application/json": {"example": {"detail": "Bearer authentication is needed"}}}, "description": "Unauthorized"}, "404": {"content": {"application/json": {"example": {"detail": "User cannot be found"}}}, "description": "Not Found"}, "500": {"description": "Internal Server Error"}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/v1/users/": {"post": {"tags": ["Users"], "summary": "Post User", "description": "Create a user account.", "operationId": "post_user_v1_users__post", "security": [{"HTTPBearer": []}], "requestBody": {"required": true, "content": {"application/json": {"schema": {"$ref": "#/components/schemas/CreateUserAccount"}}}}, "responses": {"201": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/UserAccount"}, "example": {"uid": "XXXXXXXXXXXXXXXXXXXXXXXXXXXX", "name": "Alan Turing", "nickname": "turingComplete", "picture": "http://t2.gstatic.com/licensed-image?q=tbn:ANd9GcQCUUTlwrAsh5cZ9yds4YIELmAgLT-kyXXJ7vhrzQLriwYb3VQqSTToKpq7heaNcMZX", "win": 0, "lose": 0, "favourite_category": "-", "max_points": 0}}}}, "401": {"content": {"application/json": {"example": {"detail": "Bearer authentication is needed"}}}, "description": "Unauthorized"}, "409": {"content": {"application/json": {"example": {"detail": "Account already exists"}}}, "description": "Conflict"}, "500": {"description": "Internal Server Error"}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}, "get": {"tags": ["Users"], "summary": "Get User", "description": "Get user details.", "operationId": "get_user_v1_users__get", "security": [{"HTTPBearer": []}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/UserAccount"}, "example": {"uid": "XXXXXXXXXXXXXXXXXXXXXXXXXXXX", "name": "Alan Turing", "nickname": "turingComplete", "picture": "http://t2.gstatic.com/licensed-image?q=tbn:ANd9GcQCUUTlwrAsh5cZ9yds4YIELmAgLT-kyXXJ7vhrzQLriwYb3VQqSTToKpq7heaNcMZX", "win": 0, "lose": 0, "favourite_category": "-", "max_points": 0}}}}, "401": {"content": {"application/json": {"example": {"detail": "Bearer authentication is needed"}}}, "description": "Unauthorized"}, "404": {"content": {"application/json": {"example": {"detail": "User cannot be found"}}}, "description": "Not Found"}, "500": {"description": "Internal Server Error"}}}}}, "components": {"schemas": {"CategoriesModel": {"title": "CategoriesModel", "required": ["categories"], "type": "object", "properties": {"categories": {"title": "Categories", "type": "array", "items": {"type": "string"}}}, "description": "A model that contains a list of categories.\n:param categories: - the list of categories"}, "CreateUserAccount": {"title": "CreateUserAccount", "required": ["nickname"], "type": "object", "properties": {"nickname": {"title": "Nickname", "maxLength": 30, "minLength": 3, "pattern": "^[A-Za-z0-9]+([A-Za-z0-9]*|[._-]?[A-Za-z0-9]+)*$", "type": "string"}}, "description": "This is used to create a new user account.\n:param nickname: - the nickname for the user account."}, "CreateUserQuiz": {"title": "CreateUserQuiz", "required": ["title", "category", "difficulty", "questions"], "type": "object", "properties": {"title": {"title": "Title", "maxLength": 200, "minLength": 1, "type": "string"}, "category": {"title": "Category", "anyOf": [{"maxLength": 30, "minLength": 5, "type": "string"}, {"type": "array", "items": {"maxLength": 30, "minLength": 5, "type": "string"}}]}, "difficulty": {"title": "Difficulty", "enum": ["easy", "medium", "hard"], "type": "string"}, "tags": {"title": "Tags", "anyOf": [{"type": "string"}, {"type": "array", "items": {"type": "string"}}]}, "questions": {"title": "Questions", "type": "array", "items": {"$ref": "#/components/schemas/Question"}}}, "additionalProperties": false, "description": "This will be used to create a quiz for a user.\n:param title: - the title of the quiz.\n:param category: - the category of the quiz.\n:param difficulty: - the difficulty of the quiz.\n:param tags: - the tags of the quiz.\n:param questions: - the questions of the quiz."}, "DeleteUserQuizzes": {"title": "DeleteUserQuizzes", "required": ["quizzes_ids"], "type": "object", "properties": {"quizzes_ids": {"title": "Quizzes Ids", "type": "array", "items": {"type": "string"}}}, "additionalProperties": false, "description": "The DeleteUserQuizzes class is used to store the quizzes_ids to delete.\n@param quizzes_ids - the quizzes ids to delete"}, "HTTPValidationError": {"title": "HTTPValidationError", "type": "object", "properties": {"detail": {"title": "Detail", "type": "array", "items": {"$ref": "#/components/schemas/ValidationError"}}}}, "Question": {"title": "Question", "required": ["question", "correct_answer", "incorrect_answers"], "type": "object", "properties": {"question": {"title": "Question", "type": "string"}, "correct_answer": {"title": "Correct Answer", "type": "string"}, "incorrect_answers": {"title": "Incorrect Answers", "type": "array", "items": {"type": "string"}}}, "additionalProperties": false, "description": "The Question class defines the question structure. "}, "TagsModel": {"title": "TagsModel", "required": ["tags"], "type": "object", "properties": {"tags": {"title": "Tags", "type": "array", "items": {"type": "string"}}}, "description": "A model that contains a list of tags.\n:param tags: - the list of tags"}, "UpdateUserQuiz": {"title": "UpdateUserQuiz", "type": "object", "properties": {"title": {"title": "Title", "maxLength": 200, "minLength": 1, "type": "string"}, "category": {"title": "Category", "anyOf": [{"maxLength": 30, "minLength": 5, "type": "string"}, {"type": "array", "items": {"maxLength": 30, "minLength": 5, "type": "string"}}]}, "difficulty": {"title": "Difficulty", "enum": ["easy", "medium", "hard"], "type": "string"}, "tags": {"title": "Tags", "anyOf": [{"type": "string"}, {"type": "array", "items": {"type": "string"}}]}, "questions": {"title": "Questions", "type": "array", "items": {"$ref": "#/components/schemas/Question"}}}, "additionalProperties": false, "description": "The UpdateUserQuiz class is used to update the user quiz.\n:param title: - the title of the quiz.\n:param category: - the category of the quiz.\n:param difficulty: - the difficulty of the quiz.\n:param tags: - the tags of the quiz.\n:param questions: - the questions of the quiz."}, "UserAccount": {"title": "UserAccount", "required": ["uid", "nickname", "name", "picture", "win", "lose", "favourite_category", "max_points"], "type": "object", "properties": {"uid": {"title": "Uid", "maxLength": 28, "minLength": 28, "type": "string"}, "nickname": {"title": "Nickname", "maxLength": 30, "minLength": 3, "pattern": "^[A-Za-z0-9]+([A-Za-z0-9]*|[._-]?[A-Za-z0-9]+)*$", "type": "string"}, "name": {"title": "Name", "type": "string"}, "picture": {"title": "Picture", "type": "string"}, "win": {"title": "Win", "type": "integer"}, "lose": {"title": "Lose", "type": "integer"}, "favourite_category": {"title": "Favourite Category", "type": "string"}, "max_points": {"title": "Max Points", "anyOf": [{"type": "integer"}, {"type": "number"}]}}, "description": "The user account model. This is the model that is used to store the user account data.\n:param uid: - the user id, this is the unique identifier for the user account.\n:param nickname: - the nickname of the user account.\n:param name: - the name of the user account.\n:param picture: - the picture of the user account.\n:param win: - the number of wins of the user account.\n:param lose: - the number of losses of the user account.\n:param favourite_category: - the favourite category of the user account.\n:param max_points: - the maximum points of the user account."}, "UserQuiz": {"title": "UserQuiz", "required": ["uid", "title", "category", "difficulty", "questions"], "type": "object", "properties": {"uid": {"title": "Uid", "maxLength": 28, "minLength": 28, "type": "string"}, "title": {"title": "Title", "maxLength": 200, "minLength": 1, "type": "string"}, "category": {"title": "Category", "anyOf": [{"maxLength": 30, "minLength": 5, "type": "string"}, {"type": "array", "items": {"maxLength": 30, "minLength": 5, "type": "string"}}]}, "difficulty": {"title": "Difficulty", "enum": ["easy", "medium", "hard"], "type": "string"}, "tags": {"title": "Tags", "anyOf": [{"type": "string"}, {"type": "array", "items": {"type": "string"}}]}, "questions": {"title": "Questions", "type": "array", "items": {"$ref": "#/components/schemas/Question"}}}, "description": "The UserQuiz class is used to create a model that can be used to create a quiz.\n:param uid: - the quiz id\n:param title: - the quiz title\n:param category: - the quiz category\n:param difficulty: - the quiz difficulty\n:param tags: - the quiz tags\n:param questions: - the questions in the quiz"}, "UserQuizzes": {"title": "UserQuizzes", "required": ["quiz_id"], "type": "object", "properties": {"quiz_id": {"$ref": "#/components/schemas/UserQuiz"}}, "description": "The UserQuizzes class is used to store the quiz_id.\n:param quiz_id: - the quiz_id of the quiz."}, "ValidationError": {"title": "ValidationError", "required": ["loc", "msg", "type"], "type": "object", "properties": {"loc": {"title": "Location", "type": "array", "items": {"anyOf": [{"type": "string"}, {"type": "integer"}]}}, "msg": {"title": "Message", "type": "string"}, "type": {"title": "Error Type", "type": "string"}}}}, "securitySchemes": {"HTTPBearer": {"type": "http", "scheme": "bearer"}}}}